Minor refactor of the parser
Renaming a function so it has "Builderfile" instead of "Bobfile" in the
name.  Admittedly this is a bit confusing.  The reason is that the
package and files related to bobfiles must be named something else so
they don't conflict with the Bobfile at the top level.  The result is
many instances in code where a "Bobfile" is referred to as a
"Builderfile".  That being the case, we should probably keep it
consistent throughout the code.

Also removing an unneeded logger global for the parser package