Refactoring build invocation to use a buildconfig
so we can specify both the bobfile path and the top.  The top is
determined internally by docker-builder, so it may contain symlinks.
Only the relative path from top to the bobfile may not contain symlinks.
Without this change, if the workdir used by the server contained a
symlink, the build would be failed